Se agrego la redireccion del login a mostrador con usuarios con rol cajero y se protegieron las rutas de mostrador

+++ app/models/detallesTicket.py
@@ -5,5 +5,5 @@ class DetalleTicket(db.Model):
     idDetalle=db.Column(db.Integer, primary_key=True)
     idTicket=db.Column(db.Integer, db.ForeignKey('ticket.idTicket'), nullable=False)
     idProducto=db.Column(db.Integer, db.ForeignKey('producto.idProducto'), nullable=False)
-    cantidad=db.Column(db.Float, nullable=False)
+    cantidad=db.Column(db.Integer, nullable=False)
     subtotal=db.Column(db.Float, nullable=False)
